created particle_initializer class
